Update and rename 02.mariohw.c to 02.lasttry1.c

+++ G/03/18/02.lasttry1.c
@@ -11,7 +11,7 @@ i++;
 avr=avr+m%10;
 m=m/10;
 }while(m>0);
-avr=avr/i;
+avr=(float)avr/i;
 if(avr>7){
 printf("heavy");
 }
